requirements.txt
# requirements.txt

numpy>=1.20.0
pandas>=1.3.0
scanpy>=1.8.0
anndata>=0.8.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pyyaml>=6.0
rpy2>=3.4.0